fix: moved to enrollmentpage
